Changed the way run, kill, and func work on start
Added multiple new methods to func.ps1, some of which are
not utilized at this point. Moved most of the code in kill to func.
added a if/else check in run that will theoretically prevent multiple
instances of run from being opened at once. This will only work if
the instances are run in their own process using either the shortcuts
or Start-Process ./scripts/run.ps1. If they are opened directly then
the name of the process will not match the check and it the safe guard
will fail.